% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ann.data}
\alias{ann.data}
\title{Chlamydia pneumoniae genome annotation.}
\format{
A data frame with 2218 rows and 9 variables:
\describe{
  \item{seqid}{sequence identification number}
  \item{source}{source database name}
  \item{type}{type of annotated region}
  \item{start}{region's start position}
  \item{end}{region's end position}
  \item{score}{score}
  \item{strand}{strand}
  \item{phase}{phase}
  \item{attribute}{region description}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/}
}
\usage{
ann.data
}
\description{
A dataset containing Chlamydia pneumoniae TW-183 (complete sequence, NC_005043.1.) genome annotation
}
\keyword{datasets}
