% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{cut_string}
\alias{cut_string}
\title{Cut string into segments}
\usage{
cut_string(string, size)
}
\arguments{
\item{string}{character string; vector of length 1}

\item{size}{integral; vector of length of segments}
}
\value{
Data frame with segment size, start and end point, segment string.
}
\description{
Cuts character string into segments of given size.
}
\details{
This function works with one string only.
Segments are cut from start to end of a string.
\code{size} might be a vector of any length, all possible variants will be cut.
}
\examples{
cut_string (string = "aaatttttttccgc", size = 12:14)

}
\author{
Elena N. Filatova
}
