% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{make_ids}
\alias{make_ids}
\title{Create unique identification values}
\usage{
make_ids(var, sep = "_")
}
\arguments{
\item{var}{vector of values}

\item{sep}{character; string to separate the terms}
}
\value{
Character vector of \code{var} with attached numbers.
}
\description{
Creates unique identification values by adding numbers to identical values.
}
\details{
This function takes vector with same values and adds numbers to create unique values.
}
\examples{
var<-c("one", "two", "three", "one", "two", "three", "one")
make_ids(var)

}
\author{
Elena N. Filatova
}
