% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta.all}
\alias{meta.all}
\title{Metadata of all available Chlamydia pneumoniae's sequences.}
\format{
A data frame with 9062 rows and 21 variables:
\describe{
  \item{uid}{sequence identification number}
  \item{gi}{sequence identification number}
  \item{GB_AcNum}{sequence identification number}
  \item{createdate}{date of note's creation}
  \item{updatedate}{date of last note's update}
  \item{source_db}{database}
  \item{organism}{organism name}
  \item{title}{sequence title}
  \item{strain}{strain}
  \item{taxid}{taxon identificator}
  \item{length}{sequence length}
  \item{biomol}{biomolecule}
  \item{moltype}{molecular type}
  \item{genome}{genome type}
  \item{complete}{sequence completness}
  \item{geneticcode}{type of genetic and codon codes}
  \item{strand}{strand}
  \item{host}{host}
  \item{country}{country}
  \item{isolation_source}{isolation material}
  \item{collection_date}{collection date}
  }
}
\source{
\url{https://www.ncbi.nlm.nih.gov/}
}
\usage{
meta.all
}
\description{
A dataset containing metadata of all Chlamydia pneumoniae's nucleotide sequences
that were downloaded from NCBI Nucleotide database (November, 2021)
}
\keyword{datasets}
